requirements.txt
psycopg2-binary==2.9.11
pandas==2.3.1
matplotlib==3.10.5
seaborn==0.13.2
numpy==2.3.2